fix: few improvements after PR review
